editing: Add "add_edit_note()" function
Adds an edit note to the last (or very recently) made edit. This is
necessary e.g. for ISRC submission via web service, as it has no support
for edit notes. The "identify" argument is a function
    function(str, str) -> bool
which receives the edit number as first, the raw html body of the edit
as second argument, and determines if the note should be added to this
edit.